mysite/requirements.txt
asgiref==3.2.10
Django==3.0.8
django-autocomplete-light==3.5.1
django-jsonfield==1.4.0
django-picklefield==3.0.1
django-pwa==1.0.10
django-tinymce==3.0.2
django-uuidfield-2==0.6.6
django-widget-tweaks==1.4.8
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
